pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-incident-response-env"
version = "0.1.0"
description = "Simulated incident-response environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.0",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn[standard]>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "requests>=2.31.0",
]

[project.scripts]
server = "incident_response_env.server.app:main"

[tool.setuptools]
packages = ["incident_response_env", "incident_response_env.server"]
package-dir = {"incident_response_env" = ".", "incident_response_env.server" = "server"}
